initial draft
Goal hints are added as an alternative to way of the hero. Goals are defined as obtaining a set of items and/or having access to a location. Default goals for dungeon rewards, skull tokens, and Triforce pieces are added (and able to be overridden via hint dist or plando). Location goals are not completed yet.